+++ zarrs/src/array/array_async_readable.rs
@@ -16,12 +16,12 @@ use super::{
     array_bytes::{copy_fill_value_into, merge_chunks_vlen},
     codec::{
         ArrayToBytesCodecTraits, AsyncArrayPartialDecoderTraits, AsyncStoragePartialDecoder,
-        CodecOptions,
+        CodecError, CodecOptions,
     },
     concurrency::concurrency_chunks_and_codec,
     element::ElementOwned,
     Array, ArrayBytes, ArrayBytesFixedDisjointView, ArrayCreateError, ArrayError, ArrayMetadata,
-    ArrayMetadataV2, ArrayMetadataV3, ArraySize, DataTypeSize,
+    ArrayMetadataV2, ArrayMetadataV3, DataTypeSize,
 };
 
 #[cfg(feature = "ndarray")]
@@ -342,9 +342,13 @@ impl<TStorage: ?Sized + AsyncReadableStorageTraits + 'static> Array<TStorage> {
             Ok(chunk)
         } else {
             let chunk_shape = self.chunk_shape(chunk_indices)?;
-            let array_size =
-                ArraySize::new(self.data_type().size(), chunk_shape.num_elements_u64());
-            Ok(ArrayBytes::new_fill_value(array_size, self.fill_value()))
+            ArrayBytes::new_fill_value(
+                self.data_type(),
+                chunk_shape.num_elements_u64(),
+                self.fill_value(),
+            )
+            .map_err(CodecError::from)
+            .map_err(ArrayError::from)
         }
     }
 
@@ -662,11 +666,13 @@ impl<TStorage: ?Sized + AsyncReadableStorageTraits + 'static> Array<TStorage> {
         // Retrieve chunk bytes
         let num_chunks = chunks.num_elements_usize();
         match num_chunks {
-            0 => {
-                let array_size =
-                    ArraySize::new(self.data_type().size(), array_subset.num_elements());
-                Ok(ArrayBytes::new_fill_value(array_size, self.fill_value()))
-            }
+            0 => ArrayBytes::new_fill_value(
+                self.data_type(),
+                array_subset.num_elements(),
+                self.fill_value(),
+            )
+            .map_err(CodecError::from)
+            .map_err(ArrayError::from),
             1 => {
                 let chunk_indices = chunks.start();
                 let chunk_subset = self.chunk_subset(chunk_indices)?;